+++ src/cli/commands/exec.js
@@ -60,7 +60,8 @@ exports.handler = (opts) => {
 
   const options = {
     cwd: opts.localPath,
-    stdio: opts.silent ? 'ignore' : 'inherit'
+    stdio: opts.silent ? 'ignore' : 'inherit',
+    shell: true
   };
 
   childProcess